+++ src/main/kotlin/lottery/domain/LottoResult.kt
@@ -1,5 +1,6 @@
 package lottery.domain
 
+import lottery.domain.rank.Rank
 import java.math.BigDecimal
 
 data class LottoResult(
