+++ src/main/kotlin/lottery/domain/rank/RankingMetric.kt
@@ -1,0 +1,5 @@
+package lottery.domain.rank
+
+fun interface RankingMetric {
+    fun rank(matchCount: Int, isBonus: Boolean): Boolean
+}
